+++ extra/scripts/phase-5b.docker.ps1
@@ -1,52 +1,66 @@
 # Enable docker
 param (
-  [string] $requiredVersion = "19.03.1",
+  [string] $requiredVersion = "19.03.5",
   [boolean] $installCompose = $True,
+  [string] $composeVersion = "",
   [string] $dockerLocation = "",
   [string] $configDockerLocation = "C:\ProgramData\Docker\config"
 )
 
-Write-Output "Phase 5 [START] - Start Phase 5b-Docker"
-Write-Output "Phase 5 [INFO] - Required Version of docker is: $requiredVersion"
-Write-Output "Phase 5 [INFO] - Install compose is set to: $installCompose"
-Write-Output "Phase 5 [INFO] - Docker location is set to: $dockerLocation"
+Write-Output "Phase 5b [START] - Start Phase 5b-Docker"
+[Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
+Write-Output "Phase 5b [INFO] - Required Version of docker is : $requiredVersion"
+Write-Output "Phase 5b [INFO] - Install compose is set to     : $installCompose"
+Write-Output "Phase 5b [INFO] - Docker compose is set to      : $composeVersion"
+Write-Output "Phase 5b [INFO] - Docker location is set to     : $dockerLocation"
 
 try {
-  Write-Output "Phase 5b-docker - Install Dockermsftprovider"
-  Install-Module -Name DockerMsftProvider -Repository PSGallery -Confirm:$false -Verbose -Force
-  Update-Module DockerMsftProvider -Force -Confirm:$false -Verbose
+  Write-Output "Phase 5b [INFO] - Install Dockermsftprovider"
+  Install-Module -Name DockerMsftProvider -Repository PSGallery -Confirm:$false -Force
+  Update-Module DockerMsftProvider -Force -Confirm:$false
 }
 catch {
-  Write "Phase 5b-docker failed - Install/Update module problem"
-  exit (-1)
+  Write "Phase 5b [ERROR] - Install/Update module problem"
+  exit (1)
 }
-
+Write-Output "Phase 5b [INFO] - Install Dockermsftprovider succesful"
 try {
-  Write-Output "Phase 5b-docker - Install Docker Package"
-  Install-Package -Name docker -ProviderName DockerMsftProvider -RequiredVersion $RequiredVersion -Confirm:$false -Verbose -Force
+  Write-Output "Phase 5b [INFO] - Install Docker package"
+  # display all versions
+  Write-Output "Phase 5b [INFO] - Display all Docker packages"
+  Find-Package -providerName DockerMsftProvider -AllVersions
+  # install version
+  Install-Package -Name docker -ProviderName DockerMsftProvider -RequiredVersion $RequiredVersion -Confirm:$false -Force
   Set-Service -Name docker -StartupType Automatic
 }
 catch {
-  Write "Phase 5b-docker failed - Install-Package Docker problem"
-  exit (-1)
+  Write "Phase 5b [ERROR] - Install-Package Docker problem"
+  exit (1)
 }
-
+Write-Output "Phase 5b [INFO] - Install Docker package successful"
 
 if ($installCompose) {
   try {
-    choco install docker-compose -y --no-progress --limit-output
+    if ($composeVersion) {
+      choco install docker-compose -y --version $composeVersion --no-progress --limit-output
+    }
+    else {
+      choco install docker-compose -y --no-progress --limit-output
+    }
+
   }
   catch {
-    Write "Phase 5b-docker failed - Install docker-compose problem"
+    Write "Phase 5b [ERROR] - Install docker-compose problem"
+    exit (1)
   }
 }
 
 If (! $dockerLocation) {
-  Write-Output "Phase 5b-docker Not changing default docker location"
+  Write-Output "Phase 5b [INFO] - Not changing default docker location"
 }
 else {
   try {
-      Write-Output "Phase 5b-docker - changing default docker location"
+      Write-Output "Phase 5b [INFO] - changing default docker location"
       Stop-Service docker -Force
       $dataRoot=@{ "data-root"="$dockerLocation" }
       if (! (Test-Path -Path $configDockerLocation )) {
@@ -56,7 +70,9 @@ else {
       Write-Output $dataRoot|ConvertTo-Json|Set-Content $configDockerLocation\daemon.json
       Start-Service docker
   }
-  catch {}
+  catch {
+
+  }
 }
-Write-Output "End Phase 5b"
+Write-Output "Phase 5b [END] - End Phase 5b"